+++ app/Kafka/Exception/TopicNotFoundException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Kafka\Exception;
+
+use Exception;
+
+class TopicNotFoundException extends Exception
+{
+}