+++ config/kafka.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    'topics' => [
+        'notification.user' => [
+            'acks' => 'all',
+        ],
+    ],
+];